requirements.txt
Flask
numpy
scipy
Pillow
